@@ -1,22 +1,20 @@
-import 'dart:math' as math;
-
 import 'package:davi/src/column_width_behavior.dart';
-import 'package:davi/src/internal/row_callbacks.dart';
+import 'package:davi/src/internal/column_notifier.dart';
+import 'package:davi/src/internal/davi_context.dart';
+import 'package:davi/src/internal/hover_notifier.dart';
 import 'package:davi/src/internal/scroll_controllers.dart';
 import 'package:davi/src/internal/table_layout_builder.dart';
 import 'package:davi/src/internal/theme_metrics/theme_metrics.dart';
-import 'package:davi/src/last_row_widget_listener.dart';
+import 'package:davi/src/trailing_widget_listener.dart';
 import 'package:davi/src/last_visible_row_listener.dart';
 import 'package:davi/src/model.dart';
 import 'package:davi/src/row_callback_typedefs.dart';
 import 'package:davi/src/row_color.dart';
-import 'package:davi/src/row_cursor.dart';
+import 'package:davi/src/row_cursor_builder.dart';
 import 'package:davi/src/row_hover_listener.dart';
 import 'package:davi/src/theme/theme.dart';
 import 'package:davi/src/theme/theme_data.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 /// Table view designed for a large number of data.
 ///
@@ -25,10 +23,10 @@ class Davi<DATA> extends StatefulWidget {
 //TODO handle negative values
 //TODO allow null and use defaults?
   const Davi(this.model,
-      {Key? key,
+      {super.key,
       this.onHover,
       this.unpinnedHorizontalScrollController,
-      this.pinnedHorizontalScrollController,
+      this.leftPinnedHorizontalScrollController,
       this.verticalScrollController,
       this.onLastVisibleRow,
       this.onRowTap,
@@ -38,36 +36,81 @@ class Davi<DATA> extends StatefulWidget {
       this.columnWidthBehavior = ColumnWidthBehavior.scrollable,
       int? visibleRowsCount,
       this.focusable = true,
-      this.tapToSortEnabled = true,
-      this.lastRowWidget,
+      this.trailingWidget,
+      this.placeholderWidget,
       this.rowColor,
       this.rowCursor,
-      this.onLastRowWidget})
+      this.semanticsEnabled = false,
+      this.onTrailingWidget})
       : visibleRowsCount = visibleRowsCount == null || visibleRowsCount > 0
             ? visibleRowsCount
-            : null,
-        super(key: key);
+            : null;
 
-  final DaviModel<DATA>? model;
+  /// The data model.
+  final DaviModel<DATA> model;
+
+  /// The horizontal scroll controller for the unpinned area of the table.
+  /// It controls the scrolling behavior of the section that is unpinned.
   final ScrollController? unpinnedHorizontalScrollController;
-  final ScrollController? pinnedHorizontalScrollController;
-  final ScrollController? verticalScrollController;
-  final OnRowHoverListener? onHover;
-  final DaviRowColor<DATA>? rowColor;
-  final DaviRowCursor<DATA>? rowCursor;
-  final RowDoubleTapCallback<DATA>? onRowDoubleTap;
-  final RowTapCallback<DATA>? onRowTap;
-  final RowTapCallback<DATA>? onRowSecondaryTap;
-  final RowTapUpCallback<DATA>? onRowSecondaryTapUp;
-  final ColumnWidthBehavior columnWidthBehavior;
-  final int? visibleRowsCount;
-  final OnLastVisibleRowListener? onLastVisibleRow;
-  final bool focusable;
-  final Widget? lastRowWidget;
-  final OnLastRowWidgetListener? onLastRowWidget;
 
-  /// Indicates whether sorting events are enabled on the header.
-  final bool tapToSortEnabled;
+  /// The horizontal scroll controller for the left pinned area of the table.
+  /// It controls the scrolling behavior of the section that is pinned to the left.
+  final ScrollController? leftPinnedHorizontalScrollController;
+
+  /// The vertical scroll controller for the table, allowing programmatic control of vertical scrolling.
+  final ScrollController? verticalScrollController;
+
+  /// A callback that is triggered when a row is hovered over.
+  final OnRowHoverListener? onHover;
+
+  /// A callback that defines the row color based on the row data.
+  final DaviRowColor<DATA>? rowColor;
+
+  /// A callback to build a custom cursor when hovering over a row.
+  final RowCursorBuilder<DATA>? rowCursor;
+
+  /// A callback that is triggered when a row is double-tapped.
+  final RowDoubleTapCallback<DATA>? onRowDoubleTap;
+
+  /// A callback that is triggered when a row is tapped.
+  final RowTapCallback<DATA>? onRowTap;
+
+  /// A callback that is triggered when a row receives a secondary tap (usually right-click).
+  final RowTapCallback<DATA>? onRowSecondaryTap;
+
+  /// A callback that is triggered when a secondary tap (usually right-click) is released over a row.
+  final RowTapUpCallback<DATA>? onRowSecondaryTapUp;
+
+  /// Defines column width behavior.
+  final ColumnWidthBehavior columnWidthBehavior;
+
+  /// The number of visible rows currently displayed in the table.
+  /// It is particularly useful when the table has an unbounded height,
+  /// as it helps determine the number of rows currently visible in the view.
+  final int? visibleRowsCount;
+
+  /// A callback that is triggered when the last visible row in the table is rendered.
+  /// This can be used to perform actions when the table reaches its last visible row.
+  final LastVisibleRowListener? onLastVisibleRow;
+
+  /// Defines whether the component is focusable.
+  final bool focusable;
+
+  /// An optional widget displayed at the end of the table's content.
+  final Widget? trailingWidget;
+
+  /// A callback that is triggered when the trailing widget appears in the table.
+  final TrailingWidgetListener? onTrailingWidget;
+
+  /// Activates semantics by adding a Semantics widget internally,
+  /// but it may degrade performance.
+  final bool semanticsEnabled;
+
+  /// An optional widget that replaces the body of the table
+  /// with a temporary visual element. It can be used to display a custom
+  /// state or visual indication, such as during
+  /// loading or other transitional states.
+  final Widget? placeholderWidget;
 
   @override
   State<StatefulWidget> createState() => _DaviState<DATA>();
@@ -76,38 +119,31 @@ class Davi<DATA> extends StatefulWidget {
 /// The [Davi] state.
 class _DaviState<DATA> extends State<Davi<DATA>> {
   late ScrollControllers _scrollControllers;
+  late Listenable _listenable;
   bool _scrolling = false;
-  int? _hoveredRowIndex;
   bool _lastRowWidgetVisible = false;
   int? _lastVisibleRow;
-  final FocusNode _focusNode = FocusNode(debugLabel: 'Davi');
+  final HoverNotifier _hoverNotifier = HoverNotifier();
+  final ColumnNotifier _columnNotifier = ColumnNotifier();
 
-  void _setHoveredRowIndex(int? rowIndex) {
-    if (widget.model != null && _hoveredRowIndex != rowIndex) {
-      _hoveredRowIndex = rowIndex;
-      if (widget.onHover != null) {
-        widget.onHover!(_hoveredRowIndex);
-      }
-    }
-  }
+  final FocusNode _focusNode = FocusNode(debugLabel: 'Davi');
 
   @override
   void initState() {
     super.initState();
     _scrollControllers = ScrollControllers(
         unpinnedHorizontal: widget.unpinnedHorizontalScrollController,
-        leftPinnedHorizontal: widget.pinnedHorizontalScrollController,
+        leftPinnedHorizontal: widget.leftPinnedHorizontalScrollController,
         vertical: widget.verticalScrollController);
-    _scrollControllers.unpinnedHorizontal.addListener(_rebuild);
-    _scrollControllers.leftPinnedHorizontal.addListener(_rebuild);
-    widget.model?.addListener(_rebuild);
+    _hoverNotifier.addListener(_onHover);
+    _buildListenable();
   }
 
   @override
   void dispose() {
-    _scrollControllers.unpinnedHorizontal.removeListener(_rebuild);
-    _scrollControllers.leftPinnedHorizontal.removeListener(_rebuild);
-    widget.model?.removeListener(_rebuild);
+    _hoverNotifier.removeListener(_onHover);
+    _hoverNotifier.dispose();
+    _columnNotifier.dispose();
     _focusNode.dispose();
     super.dispose();
   }
@@ -115,29 +151,16 @@ class _DaviState<DATA> extends State<Davi<DATA>> {
   @override
   void didUpdateWidget(covariant Davi<DATA> oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (widget.verticalScrollController != null &&
-        _scrollControllers.vertical != widget.verticalScrollController) {
-      _scrollControllers.vertical = widget.verticalScrollController!;
-    }
-    if (widget.unpinnedHorizontalScrollController != null &&
-        _scrollControllers.unpinnedHorizontal !=
-            widget.unpinnedHorizontalScrollController) {
-      _scrollControllers.unpinnedHorizontal.removeListener(_rebuild);
-      _scrollControllers.unpinnedHorizontal =
-          widget.unpinnedHorizontalScrollController!;
-      _scrollControllers.unpinnedHorizontal.addListener(_rebuild);
-    }
-    if (widget.pinnedHorizontalScrollController != null &&
-        _scrollControllers.leftPinnedHorizontal !=
-            widget.pinnedHorizontalScrollController) {
-      _scrollControllers.leftPinnedHorizontal.removeListener(_rebuild);
-      _scrollControllers.leftPinnedHorizontal =
-          widget.pinnedHorizontalScrollController!;
-      _scrollControllers.leftPinnedHorizontal.addListener(_rebuild);
+    if (_scrollControllers.update(
+        unpinnedHorizontal: widget.unpinnedHorizontalScrollController,
+        leftPinnedHorizontal: widget.leftPinnedHorizontalScrollController,
+        vertical: widget.verticalScrollController)) {
+      setState(() {
+        // rebuild subtree with the new scroll controllers.
+      });
     }
     if (widget.model != oldWidget.model) {
-      oldWidget.model?.removeListener(_rebuild);
-      widget.model?.addListener(_rebuild);
+      _buildListenable();
       if (_scrollControllers.vertical.hasClients) {
         _scrollControllers.vertical.jumpTo(0);
       }
@@ -150,11 +173,107 @@ class _DaviState<DATA> extends State<Davi<DATA>> {
     }
   }
 
-  void _onLastRowWidget(bool visible) {
-    if (widget.onLastRowWidget != null) {
+  void _buildListenable() {
+    _listenable = Listenable.merge([widget.model, _columnNotifier]);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final DaviThemeData theme = DaviTheme.of(context);
+    if (theme.cell.overrideInputDecoration) {
+      return Theme(
+          data: ThemeData(
+              inputDecorationTheme: const InputDecorationTheme(
+                  isDense: true, border: InputBorder.none)),
+          child: _decoratedContainer(context));
+    }
+    return _decoratedContainer(context);
+  }
+
+  Widget _decoratedContainer(BuildContext context) {
+    final DaviThemeData theme = DaviTheme.of(context);
+    if (theme.decoration != null) {
+      return Container(
+          decoration: theme.decoration, child: _placeholder(context));
+    }
+    return _placeholder(context);
+  }
+
+  Widget _placeholder(BuildContext context) {
+    if (widget.placeholderWidget != null) {
+      return widget.placeholderWidget!;
+    }
+    return _cursorBugWorkaround(context);
+  }
+
+  Widget _cursorBugWorkaround(BuildContext context) {
+    final DaviThemeData theme = DaviTheme.of(context);
+    if (theme.decoration?.color != null) {
+      return _listenableBuilder();
+    }
+    // To avoid the bug that makes a cursor disappear
+    // (https://github.com/flutter/flutter/issues/106767),
+    // always build a Container with some color.
+    return Container(color: Colors.transparent, child: _listenableBuilder());
+  }
+
+  Widget _listenableBuilder() {
+    return ListenableBuilder(listenable: _listenable, builder: _builder);
+  }
+
+  Widget _builder(BuildContext context, Widget? child) {
+    final DaviThemeData theme = DaviTheme.of(context);
+    final TableThemeMetrics themeMetrics = TableThemeMetrics(theme);
+
+    final DaviContext<DATA> daviContext = DaviContext(
+        hoverNotifier: _hoverNotifier,
+        hasHoverListener: widget.onHover != null,
+        columnNotifier: _columnNotifier,
+        semanticsEnabled: widget.semanticsEnabled,
+        model: widget.model,
+        onLastVisibleRow: _onLastVisibleRowListener,
+        onTrailingWidget: _onTrailingWidget,
+        rowColor: widget.rowColor,
+        focusable: widget.focusable,
+        focusNode: _focusNode,
+        rowCursorBuilder: widget.rowCursor,
+        trailingWidget: widget.trailingWidget,
+        onRowTap: widget.onRowTap,
+        onRowSecondaryTap: widget.onRowSecondaryTap,
+        onRowSecondaryTapUp: widget.onRowSecondaryTapUp,
+        onRowDoubleTap: widget.onRowDoubleTap,
+        scrolling: _scrolling,
+        visibleRowsCount: widget.visibleRowsCount,
+        columnWidthBehavior: widget.columnWidthBehavior,
+        themeMetrics: themeMetrics,
+        scrollControllers: _scrollControllers);
+
+    return FocusTraversalGroup(
+        policy: _NoTraversalPolicy(),
+        child: Listener(
+          behavior: HitTestBehavior.translucent,
+          onPointerDown: (pointer) {
+            if (widget.model.isRowsNotEmpty && widget.focusable) {
+              _focusNode.requestFocus();
+            }
+          },
+          child: ClipRect(
+              child: TableLayoutBuilder(
+                  daviContext: daviContext, onDragScroll: _onDragScroll)),
+        ));
+  }
+
+  void _onHover() {
+    if (widget.onHover != null) {
+      widget.onHover!(_hoverNotifier.index);
+    }
+  }
+
+  void _onTrailingWidget(bool visible) {
+    if (widget.onTrailingWidget != null) {
       if (_lastRowWidgetVisible != visible) {
         _lastRowWidgetVisible = visible;
-        Future.microtask(() => widget.onLastRowWidget!(_lastRowWidgetVisible));
+        Future.microtask(() => widget.onTrailingWidget!(_lastRowWidgetVisible));
       }
     }
   }
@@ -168,131 +287,24 @@ class _DaviState<DATA> extends State<Davi<DATA>> {
     }
   }
 
-  void _rebuild() {
-    setState(() {});
+  void _onDragScroll(bool running) {
+    _hoverNotifier.enabled = !running;
+    setState(() => _scrolling = running);
   }
+}
+
+class _NoTraversalPolicy extends FocusTraversalPolicy {
+  @override
+  FocusNode? findFirstFocusInDirection(
+          FocusNode currentNode, TraversalDirection direction) =>
+      null;
 
   @override
-  Widget build(BuildContext context) {
-    final DaviThemeData theme = DaviTheme.of(context);
+  bool inDirection(FocusNode currentNode, TraversalDirection direction) =>
+      false;
 
-    final TableThemeMetrics themeMetrics = TableThemeMetrics(theme);
-
-    Widget table = ClipRect(
-        child: TableLayoutBuilder(
-            onHover: widget.onHover != null ? _setHoveredRowIndex : null,
-            tapToSortEnabled: widget.tapToSortEnabled,
-            scrollControllers: _scrollControllers,
-            columnWidthBehavior: widget.columnWidthBehavior,
-            themeMetrics: themeMetrics,
-            visibleRowsLength: widget.visibleRowsCount,
-            onLastRowWidget: _onLastRowWidget,
-            onLastVisibleRow: _onLastVisibleRowListener,
-            model: widget.model,
-            scrolling: _scrolling,
-            rowColor: widget.rowColor,
-            rowCursor: widget.rowCursor,
-            lastRowWidget: widget.lastRowWidget,
-            rowCallbacks: RowCallbacks(
-                onRowTap: widget.onRowTap,
-                onRowSecondaryTap: widget.onRowSecondaryTap,
-                onRowSecondaryTapUp: widget.onRowSecondaryTapUp,
-                onRowDoubleTap: widget.onRowDoubleTap),
-            onDragScroll: _onDragScroll));
-
-    if (widget.model != null) {
-      if (theme.row.hoverBackground != null ||
-          theme.row.hoverForeground != null) {
-        table = MouseRegion(
-            onExit: (event) => _setHoveredRowIndex(null), child: table);
-      }
-
-      table = Listener(
-        behavior: HitTestBehavior.translucent,
-        onPointerDown: (pointer) {
-          if (widget.focusable) {
-            _focusNode.requestFocus();
-          }
-        },
-        onPointerSignal: (pointerSignal) {
-          if (pointerSignal is PointerScrollEvent &&
-              _scrollControllers.vertical.hasClients &&
-              pointerSignal.scrollDelta.dy != 0) {
-            _scrollControllers.vertical.position
-                .pointerScroll(pointerSignal.scrollDelta.dy);
-          }
-        },
-        onPointerPanZoomUpdate: (PointerPanZoomUpdateEvent event) {
-          // trackpad on macOS
-          if (_scrollControllers.vertical.hasClients &&
-              event.panDelta.dy != 0) {
-            _scrollControllers.vertical.position
-                .pointerScroll(-event.panDelta.dy);
-          }
-        },
-        child: table,
-      );
-
-      if (widget.focusable) {
-        table = Focus(
-            focusNode: _focusNode,
-            onKeyEvent: (node, event) =>
-                _handleKeyPress(node, event, themeMetrics.row.height),
-            child: table);
-      }
-    }
-
-    if (theme.decoration != null) {
-      table = Container(decoration: theme.decoration, child: table);
-    }
-
-    if (theme.cell.overrideInputDecoration) {
-      table = Theme(
-          data: ThemeData(
-              inputDecorationTheme: const InputDecorationTheme(
-                  isDense: true, border: InputBorder.none)),
-          child: table);
-    }
-
-    return table;
-  }
-
-  KeyEventResult _handleKeyPress(
-      FocusNode node, KeyEvent event, double rowHeight) {
-    if (event is KeyUpEvent) {
-      if (_scrollControllers.vertical.hasClients) {
-        if (event.logicalKey == LogicalKeyboardKey.arrowDown) {
-          double target = math.min(
-              _scrollControllers.vertical.position.pixels + rowHeight,
-              _scrollControllers.vertical.position.maxScrollExtent);
-          _scrollControllers.vertical.animateTo(target,
-              duration: const Duration(milliseconds: 30), curve: Curves.ease);
-        } else if (event.logicalKey == LogicalKeyboardKey.arrowUp) {
-          double target = math.max(
-              _scrollControllers.vertical.position.pixels - rowHeight, 0);
-          _scrollControllers.vertical.animateTo(target,
-              duration: const Duration(milliseconds: 30), curve: Curves.ease);
-        } else if (event.logicalKey == LogicalKeyboardKey.pageDown) {
-          double target = math.min(
-              _scrollControllers.vertical.position.pixels +
-                  _scrollControllers.vertical.position.viewportDimension,
-              _scrollControllers.vertical.position.maxScrollExtent);
-          _scrollControllers.vertical.animateTo(target,
-              duration: const Duration(milliseconds: 30), curve: Curves.ease);
-        } else if (event.logicalKey == LogicalKeyboardKey.pageUp) {
-          double target = math.max(
-              _scrollControllers.vertical.position.pixels -
-                  _scrollControllers.vertical.position.viewportDimension,
-              0);
-          _scrollControllers.vertical.animateTo(target,
-              duration: const Duration(milliseconds: 30), curve: Curves.ease);
-        }
-      }
-    }
-    return KeyEventResult.ignored;
-  }
-
-  void _onDragScroll(bool start) {
-    setState(() => _scrolling = start);
-  }
+  @override
+  Iterable<FocusNode> sortDescendants(
+          Iterable<FocusNode> descendants, FocusNode currentNode) =>
+      descendants;
 }
