@@ -2,9 +2,8 @@ import 'package:davi/src/column.dart';
 import 'package:davi/src/internal/columns_layout.dart';
 import 'package:davi/src/internal/columns_layout_child.dart';
 import 'package:davi/src/internal/header_cell.dart';
-import 'package:davi/src/internal/scroll_offsets.dart';
+import 'package:davi/src/internal/davi_context.dart';
 import 'package:davi/src/internal/table_layout_settings.dart';
-import 'package:davi/src/model.dart';
 import 'package:davi/src/theme/theme.dart';
 import 'package:davi/src/theme/theme_data.dart';
 import 'package:flutter/widgets.dart';
@@ -13,57 +12,52 @@ import 'package:meta/meta.dart';
 @internal
 class HeaderWidget<DATA> extends StatelessWidget {
   const HeaderWidget(
-      {Key? key,
+      {super.key,
+      required this.daviContext,
       required this.layoutSettings,
-      required this.model,
-      required this.resizable,
-      required this.horizontalScrollOffsets,
-      required this.tapToSortEnabled})
-      : super(key: key);
+      required this.resizable});
 
+  final DaviContext<DATA> daviContext;
   final TableLayoutSettings layoutSettings;
-  final DaviModel<DATA> model;
   final bool resizable;
-  final HorizontalScrollOffsets horizontalScrollOffsets;
-  final bool tapToSortEnabled;
 
   @override
   Widget build(BuildContext context) {
+    if (daviContext.model.isColumnsEmpty) {
+      return Container();
+    }
     DaviThemeData theme = DaviTheme.of(context);
 
     List<ColumnsLayoutChild<DATA>> children = [];
 
-    final isMultiSorted = model.isMultiSorted;
-
     for (int columnIndex = 0;
-        columnIndex < model.columnsLength;
+        columnIndex < daviContext.model.columnsLength;
         columnIndex++) {
-      final DaviColumn<DATA> column = model.columnAt(columnIndex);
+      final DaviColumn<DATA> column = daviContext.model.columnAt(columnIndex);
 
       final Widget cell = DaviHeaderCell<DATA>(
           key: ValueKey<int>(columnIndex),
-          model: model,
+          daviContext: daviContext,
           column: column,
           resizable: resizable,
-          tapToSortEnabled: tapToSortEnabled,
-          isMultiSorted: isMultiSorted,
           columnIndex: columnIndex);
       children.add(ColumnsLayoutChild<DATA>(index: columnIndex, child: cell));
     }
 
     Widget header = ColumnsLayout(
         layoutSettings: layoutSettings,
-        horizontalScrollOffsets: horizontalScrollOffsets,
-        paintDividerColumns: true,
+        scrollControllers: daviContext.scrollControllers,
+        columnDividerThickness: theme.columnDividerThickness,
+        columnDividerColor: theme.header.columnDividerColor,
         children: children);
 
     Color? color = theme.header.color;
     BoxBorder? border;
-    if (theme.header.bottomBorderHeight > 0 &&
+    if (theme.header.bottomBorderThickness > 0 &&
         theme.header.bottomBorderColor != null) {
       border = Border(
           bottom: BorderSide(
-              width: theme.header.bottomBorderHeight,
+              width: theme.header.bottomBorderThickness,
               color: theme.header.bottomBorderColor!));
     }
 
