@@ -1,4 +1,3 @@
-import 'package:davi/src/internal/scroll_offsets.dart';
 import 'package:davi/src/internal/table_layout_child.dart';
 import 'package:davi/src/internal/table_layout_element.dart';
 import 'package:davi/src/internal/table_layout_render_box.dart';
@@ -12,23 +11,18 @@ import 'package:meta/meta.dart';
 @internal
 class TableLayout<DATA> extends MultiChildRenderObjectWidget {
   const TableLayout(
-      {Key? key,
+      {super.key,
       required this.layoutSettings,
       required this.theme,
-      required this.horizontalScrollOffsets,
-      required List<TableLayoutChild> children})
-      : super(key: key, children: children);
+      required List<TableLayoutChild> super.children});
 
   final TableLayoutSettings layoutSettings;
   final DaviThemeData theme;
-  final HorizontalScrollOffsets horizontalScrollOffsets;
 
   @override
   RenderObject createRenderObject(BuildContext context) {
     return TableLayoutRenderBox<DATA>(
-        layoutSettings: layoutSettings,
-        theme: theme,
-        horizontalScrollOffsets: horizontalScrollOffsets);
+        layoutSettings: layoutSettings, theme: theme);
   }
 
   @override
@@ -42,7 +36,6 @@ class TableLayout<DATA> extends MultiChildRenderObjectWidget {
     super.updateRenderObject(context, renderObject);
     renderObject
       ..layoutSettings = layoutSettings
-      ..theme = theme
-      ..horizontalScrollOffsets = horizontalScrollOffsets;
+      ..theme = theme;
   }
 }
