@@ -2,7 +2,6 @@ import 'dart:math' as math;
 
 import 'package:davi/src/internal/layout_child_id.dart';
 import 'package:davi/src/internal/layout_utils.dart';
-import 'package:davi/src/internal/scroll_offsets.dart';
 import 'package:davi/src/internal/table_layout_parent_data.dart';
 import 'package:davi/src/internal/table_layout_settings.dart';
 import 'package:davi/src/theme/theme_data.dart';
@@ -16,28 +15,19 @@ class TableLayoutRenderBox<DATA> extends RenderBox
         RenderBoxContainerDefaultsMixin<RenderBox, TableLayoutParentData> {
   TableLayoutRenderBox(
       {required TableLayoutSettings layoutSettings,
-      required DaviThemeData theme,
-      required HorizontalScrollOffsets horizontalScrollOffsets})
+      required DaviThemeData theme})
       : _layoutSettings = layoutSettings,
-        _theme = theme,
-        _horizontalScrollOffsets = horizontalScrollOffsets;
+        _theme = theme;
 
   RenderBox? _header;
   RenderBox? _rows;
   RenderBox? _verticalScrollbar;
   RenderBox? _unpinnedHorizontalScrollbar;
   RenderBox? _leftPinnedHorizontalScrollbar;
-  RenderBox? _topCorner;
-  RenderBox? _bottomCorner;
-
-  HorizontalScrollOffsets _horizontalScrollOffsets;
-
-  set horizontalScrollOffsets(HorizontalScrollOffsets value) {
-    if (_horizontalScrollOffsets != value) {
-      _horizontalScrollOffsets = value;
-      markNeedsPaint();
-    }
-  }
+  RenderBox? _headerEdge;
+  RenderBox? _scrollbarEdge;
+  RenderBox? _summaryEdge;
+  RenderBox? _summary;
 
   DaviThemeData _theme;
 
@@ -73,13 +63,15 @@ class TableLayoutRenderBox<DATA> extends RenderBox
     _unpinnedHorizontalScrollbar = null;
     _leftPinnedHorizontalScrollbar = null;
     _verticalScrollbar = null;
-    _topCorner = null;
-    _bottomCorner = null;
+    _headerEdge = null;
+    _scrollbarEdge = null;
+    _summaryEdge = null;
+    _summary = null;
 
     visitChildren((child) {
       RenderBox renderBox = child as RenderBox;
       TableLayoutParentData parentData = child._parentData();
-      if (parentData.id == LayoutChildId.rows) {
+      if (parentData.id == LayoutChildId.cells) {
         _rows = renderBox;
       } else if (parentData.id == LayoutChildId.verticalScrollbar) {
         _verticalScrollbar = renderBox;
@@ -87,12 +79,16 @@ class TableLayoutRenderBox<DATA> extends RenderBox
         _unpinnedHorizontalScrollbar = renderBox;
       } else if (parentData.id == LayoutChildId.leftPinnedHorizontalScrollbar) {
         _leftPinnedHorizontalScrollbar = renderBox;
-      } else if (parentData.id == LayoutChildId.topCorner) {
-        _topCorner = renderBox;
-      } else if (parentData.id == LayoutChildId.bottomCorner) {
-        _bottomCorner = renderBox;
+      } else if (parentData.id == LayoutChildId.headerEdge) {
+        _headerEdge = renderBox;
+      } else if (parentData.id == LayoutChildId.scrollbarEdge) {
+        _scrollbarEdge = renderBox;
       } else if (parentData.id == LayoutChildId.header) {
         _header = renderBox;
+      } else if (parentData.id == LayoutChildId.summary) {
+        _summary = renderBox;
+      } else if (parentData.id == LayoutChildId.summaryEdge) {
+        _summaryEdge = renderBox;
       }
     });
 
@@ -108,6 +104,17 @@ class TableLayoutRenderBox<DATA> extends RenderBox
 
     // rows
     _layoutChild(child: _rows, bounds: _layoutSettings.cellsBounds);
+    if (_summary != null) {
+      _summary!.layout(
+          BoxConstraints.tightFor(
+              width: _layoutSettings.summaryBounds.width,
+              height: _layoutSettings.summaryBounds.height),
+          parentUsesSize: true);
+      _summary!._parentData().offset = Offset(
+          0,
+          _layoutSettings.headerBounds.height +
+              _layoutSettings.cellsBounds.height);
+    }
 
     // horizontal scrollbars
     _layoutChild(
@@ -122,26 +129,42 @@ class TableLayoutRenderBox<DATA> extends RenderBox
         child: _verticalScrollbar,
         bounds: _layoutSettings.verticalScrollbarBounds);
 
-    // top corner
-    if (_topCorner != null) {
-      _topCorner!.layout(
+    // header edge
+    if (_headerEdge != null) {
+      _headerEdge!.layout(
           BoxConstraints.tightFor(
               width: _layoutSettings.themeMetrics.scrollbar.width,
               height: _layoutSettings.themeMetrics.header.height),
           parentUsesSize: true);
-      _topCorner!._parentData().offset = Offset(
+      _headerEdge!._parentData().offset = Offset(
           constraints.maxWidth - _layoutSettings.themeMetrics.scrollbar.width,
           0);
     }
 
-    // bottom corner
-    if (_bottomCorner != null) {
-      _bottomCorner!.layout(
+    // summary edge
+    if (_summaryEdge != null) {
+      _summaryEdge!.layout(
+          BoxConstraints.tightFor(
+              width: _layoutSettings.themeMetrics.scrollbar.width,
+              height: _layoutSettings.themeMetrics.summary.height),
+          parentUsesSize: true);
+      _summaryEdge!._parentData().offset = Offset(
+          constraints.maxWidth - _layoutSettings.themeMetrics.scrollbar.width,
+          _layoutSettings.height -
+              (_layoutSettings.hasHorizontalScrollbar
+                  ? _layoutSettings.themeMetrics.scrollbar.height
+                  : 0) -
+              _layoutSettings.themeMetrics.summary.height);
+    }
+
+    // scrollbar edge
+    if (_scrollbarEdge != null) {
+      _scrollbarEdge!.layout(
           BoxConstraints.tightFor(
               width: _layoutSettings.themeMetrics.scrollbar.width,
               height: _layoutSettings.themeMetrics.scrollbar.height),
           parentUsesSize: true);
-      _bottomCorner!._parentData().offset = Offset(
+      _scrollbarEdge!._parentData().offset = Offset(
           constraints.maxWidth - _layoutSettings.themeMetrics.scrollbar.width,
           _layoutSettings.height -
               _layoutSettings.themeMetrics.scrollbar.height);
@@ -185,7 +208,7 @@ class TableLayoutRenderBox<DATA> extends RenderBox
         child: _header,
         clipBounds: _layoutSettings.headerBounds);
     _paintChild(
-        context: context, offset: offset, child: _topCorner, clipBounds: null);
+        context: context, offset: offset, child: _headerEdge, clipBounds: null);
     _paintChild(
         context: context,
         offset: offset,
@@ -204,8 +227,15 @@ class TableLayoutRenderBox<DATA> extends RenderBox
     _paintChild(
         context: context,
         offset: offset,
-        child: _bottomCorner,
+        child: _scrollbarEdge,
         clipBounds: null);
+    _paintChild(
+        context: context,
+        offset: offset,
+        child: _summaryEdge,
+        clipBounds: null);
+    _paintChild(
+        context: context, offset: offset, child: _summary, clipBounds: null);
     _paintChild(
         context: context,
         offset: offset,
@@ -227,7 +257,8 @@ class TableLayoutRenderBox<DATA> extends RenderBox
                   _layoutSettings.leftPinnedHorizontalScrollbarBounds.width,
               offset.dy +
                   _layoutSettings.headerBounds.height +
-                  _layoutSettings.cellsBounds.height,
+                  _layoutSettings.cellsBounds.height +
+                  _layoutSettings.summaryBounds.height,
               _layoutSettings.themeMetrics.columnDividerThickness,
               _layoutSettings.leftPinnedHorizontalScrollbarBounds.height),
           Paint()..color = _theme.scrollbar.columnDividerColor!);
