@@ -1,22 +1,25 @@
-library davi;
-
 export 'src/cell_background.dart';
-export 'src/cell_builder.dart';
+export 'src/cell_text_style.dart';
 export 'src/cell_icon.dart';
-export 'src/cell_style.dart';
-export 'src/column.dart';
+export 'src/column.dart' hide DaviColumnHelper;
 export 'src/column_id.dart';
 export 'src/sort.dart';
+export 'src/row_span_overflow_behavior.dart';
+export 'src/sorting_mode.dart';
+export 'src/cell_listenable_builder.dart';
 export 'src/column_width_behavior.dart';
 export 'src/davi.dart';
 export 'src/last_visible_row_listener.dart';
 export 'src/model.dart';
+export 'src/cell_collision_behavior.dart';
+export 'src/max_span_behavior.dart';
+export 'src/cell_painter.dart';
 export 'src/pin_status.dart';
 export 'src/sort_callback_typedef.dart';
-export 'src/row.dart';
 export 'src/row_callback_typedefs.dart';
 export 'src/row_color.dart';
-export 'src/row_cursor.dart';
+export 'src/summary_builder.dart';
+export 'src/row_cursor_builder.dart';
 export 'src/row_hover_listener.dart';
 export 'src/sort_direction.dart';
 export 'src/theme/cell_null_color.dart';
@@ -27,9 +30,10 @@ export 'src/theme/sort_icon.dart';
 export 'src/theme/cell_theme_data.dart';
 export 'src/theme/header_cell_theme_data.dart';
 export 'src/theme/header_theme_data.dart';
+export 'src/theme/edge_theme_data.dart';
+export 'src/theme/summary_theme_data.dart';
 export 'src/theme/row_theme_data.dart';
 export 'src/theme/scrollbar_theme_data.dart';
 export 'src/theme/theme.dart';
 export 'src/theme/theme_data.dart';
 export 'src/theme/theme_row_color.dart';
-export 'src/value_mapper.dart';
