@@ -3,6 +3,7 @@ import 'package:davi/src/internal/theme_metrics/header_cell_theme_metrics.dart';
 import 'package:davi/src/internal/theme_metrics/header_theme_metrics.dart';
 import 'package:davi/src/internal/theme_metrics/row_theme_metrics.dart';
 import 'package:davi/src/internal/theme_metrics/scrollbar_theme_metrics.dart';
+import 'package:davi/src/internal/theme_metrics/summary_theme_metrics.dart';
 import 'package:davi/src/theme/theme_data.dart';
 import 'package:meta/meta.dart';
 
@@ -19,6 +20,8 @@ class TableThemeMetrics {
         RowThemeMetrics(themeData: themeData.row, cellThemeMetrics: cell);
     TableScrollbarThemeMetrics scrollbar =
         TableScrollbarThemeMetrics(themeData: themeData.scrollbar);
+    SummaryThemeMetrics summary =
+        SummaryThemeMetrics(themeData: themeData.summary);
 
     return TableThemeMetrics._(
         columnDividerThickness: themeData.columnDividerThickness,
@@ -26,7 +29,8 @@ class TableThemeMetrics {
         header: header,
         headerCell: headerCell,
         row: row,
-        scrollbar: scrollbar);
+        scrollbar: scrollbar,
+        summary: summary);
   }
 
   TableThemeMetrics._(
@@ -35,7 +39,8 @@ class TableThemeMetrics {
       required this.header,
       required this.headerCell,
       required this.row,
-      required this.scrollbar});
+      required this.scrollbar,
+      required this.summary});
 
   final double columnDividerThickness;
   final CellThemeMetrics cell;
@@ -43,6 +48,7 @@ class TableThemeMetrics {
   final HeaderCellThemeMetrics headerCell;
   final RowThemeMetrics row;
   final TableScrollbarThemeMetrics scrollbar;
+  final SummaryThemeMetrics summary;
 
   @override
   bool operator ==(Object other) =>
@@ -54,7 +60,8 @@ class TableThemeMetrics {
           header == other.header &&
           headerCell == other.headerCell &&
           row == other.row &&
-          scrollbar == other.scrollbar;
+          scrollbar == other.scrollbar &&
+          summary == other.summary;
 
   @override
   int get hashCode =>
@@ -63,5 +70,6 @@ class TableThemeMetrics {
       header.hashCode ^
       headerCell.hashCode ^
       row.hashCode ^
-      scrollbar.hashCode;
+      scrollbar.hashCode ^
+      summary.hashCode;
 }
