@@ -1,12 +1,16 @@
 import 'dart:collection';
+import 'dart:math' as math;
 
 import 'package:collection/collection.dart';
+import 'package:davi/src/cell_collision_behavior.dart';
 import 'package:davi/src/column.dart';
+import 'package:davi/src/max_span_behavior.dart';
+import 'package:davi/src/row_span_overflow_behavior.dart';
 import 'package:davi/src/sort.dart';
 import 'package:davi/src/sort_callback_typedef.dart';
 import 'package:davi/src/sort_direction.dart';
+import 'package:davi/src/sorting_mode.dart';
 import 'package:flutter/widgets.dart';
-import 'package:meta/meta.dart';
 
 /// The [Davi] model.
 ///
@@ -16,28 +20,61 @@ class DaviModel<DATA> extends ChangeNotifier {
       {List<DATA> rows = const [],
       List<DaviColumn<DATA>> columns = const [],
       this.ignoreDataComparators = false,
-      this.alwaysSorted = false,
       this.multiSortEnabled = false,
-      this.onSort}) {
+      SortingMode sortingMode = SortingMode.interactive,
+      this.onSort,
+      int maxColumnSpan = 10,
+      int maxRowSpan = 15,
+      CellCollisionBehavior collisionBehavior = CellCollisionBehavior.ignore,
+      RowSpanOverflowBehavior rowSpanOverflowBehavior =
+          RowSpanOverflowBehavior.cap,
+      this.maxSpanBehavior = MaxSpanBehavior.throwException})
+      : maxRowSpan = math.max(maxRowSpan, 1),
+        maxColumnSpan = math.max(maxColumnSpan, 1),
+        _sortingMode = sortingMode,
+        _collisionBehavior = collisionBehavior,
+        _rowSpanOverflowBehavior = rowSpanOverflowBehavior {
     _originalRows = List.from(rows);
-    addColumns(columns);
+    _addColumns(columns, false);
     _updateRows(notify: false);
   }
 
-  late List<DATA> _rows;
   final List<DaviColumn<DATA>> _columns = [];
-  late final List<DATA> _originalRows;
 
-  /// The event that will be triggered at each sorting.
+  late final List<DATA> _originalRows;
+  late List<DATA> _sortableRows;
+  late UnmodifiableListView<DATA> _rowsView;
+  UnmodifiableListView<DATA> get rows => _rowsView;
+
+  /// The event that will be triggered before each sorting.
   OnSortCallback<DATA>? onSort;
 
   final bool multiSortEnabled;
 
+  bool _hasSummary = false;
+
+  bool get hasSummary => _hasSummary;
+
+  CellCollisionBehavior _collisionBehavior;
+
+  /// Determines the behavior when a cell collision occurs in the grid.
+  ///
+  /// This property uses [CellCollisionBehavior] to define how collisions
+  /// are handled, such as ignoring the colliding cell, logging a warning,
+  /// allowing overlap, or throwing an exception.
+  CellCollisionBehavior get collisionBehavior => _collisionBehavior;
+  set collisionBehavior(CellCollisionBehavior value) {
+    if (_collisionBehavior != value) {
+      _collisionBehavior = value;
+      notifyListeners();
+    }
+  }
+
   /// Gets the sorted columns.
   List<DaviColumn<DATA>> get sortedColumns {
     List<DaviColumn<DATA>> list =
-        _columns.where((column) => column.sort != null).toList();
-    list.sort((a, b) => a.sortPriority!.compareTo(b.sortPriority!));
+        _columns.where((column) => column.sortDirection != null).toList();
+    list.sort((a, b) => a.sortPriority.compareTo(b.sortPriority));
     return list;
   }
 
@@ -45,24 +82,30 @@ class DaviModel<DATA> extends ChangeNotifier {
   List<DaviSort> get sortList {
     List<DaviSort> list = [];
     for (DaviColumn<DATA> column in sortedColumns) {
-      final DaviSort? sort = column.sort;
-      if (sort == null) {
+      final DaviSortDirection? direction = column.sortDirection;
+      if (direction == null) {
         throw StateError('Column sort should not be null.');
       }
-      list.add(sort);
+      list.add(DaviSort(column.id, direction));
     }
     return list;
   }
 
-  /// Ignore column [dataComparator] to maintain the natural order of the data.
+  /// Specifies the sorting mode for the table: interactive, always sorted, or disabled.
+  SortingMode _sortingMode;
+  SortingMode get sortingMode => _sortingMode;
+  set sortingMode(SortingMode value) {
+    if (_sortingMode != value) {
+      _sortingMode = value;
+      _ensureSort();
+      _updateRows(notify: true);
+    }
+  }
+
+  /// Ignore column [DaviColumn.dataComparator] to maintain the natural order of the data.
   final bool ignoreDataComparators;
 
-  /// Defines if there will always be some sorted column.
-  ///
-  /// The column must be sortable.
-  final bool alwaysSorted;
-
-  bool get _isRowsModifiable => _rows is! UnmodifiableListView;
+  bool get _isRowsModifiable => _sortableRows is! UnmodifiableListView;
 
   int get originalRowsLength => _originalRows.length;
 
@@ -70,11 +113,11 @@ class DaviModel<DATA> extends ChangeNotifier {
 
   bool get isOriginalRowsNotEmpty => _originalRows.isNotEmpty;
 
-  int get rowsLength => _rows.length;
+  int get rowsLength => _sortableRows.length;
 
-  bool get isRowsEmpty => _rows.isEmpty;
+  bool get isRowsEmpty => _sortableRows.isEmpty;
 
-  bool get isRowsNotEmpty => _rows.isNotEmpty;
+  bool get isRowsNotEmpty => _sortableRows.isNotEmpty;
 
   int get columnsLength => _columns.length;
 
@@ -82,21 +125,44 @@ class DaviModel<DATA> extends ChangeNotifier {
 
   bool get isColumnsNotEmpty => _columns.isNotEmpty;
 
-  DaviColumn<DATA>? _columnInResizing;
+  /// The maximum number of rows a single cell can span.
+  ///
+  /// If a cell's `rowSpan` exceeds this value, the behavior will depend on
+  /// [maxSpanBehavior]. See [MaxSpanBehavior] for available options and details.
+  ///
+  /// Adjust this value to control the performance and usability of the grid.
+  final int maxRowSpan;
 
-  DaviColumn<DATA>? get columnInResizing => _columnInResizing;
+  /// The maximum number of columns a single cell can span.
+  ///
+  /// If a cell's `columnSpan` exceeds this value, the behavior will depend on
+  /// [maxSpanBehavior]. See [MaxSpanBehavior] for available options and details.
+  ///
+  /// Adjust this value to accommodate wider spans when necessary.
+  final int maxColumnSpan;
 
-  @internal
-  set columnInResizing(DaviColumn<DATA>? column) {
-    _columnInResizing = column;
-    notifyListeners();
+  /// Determines how to handle spans that exceed [maxRowSpan] or [maxColumnSpan].
+  ///
+  /// Refer to [MaxSpanBehavior] for details on the available policies.
+  final MaxSpanBehavior maxSpanBehavior;
+
+  /// Defines the behavior when a cell's rowSpan exceeds the available number of rows in the table.
+  RowSpanOverflowBehavior _rowSpanOverflowBehavior;
+  RowSpanOverflowBehavior get rowSpanOverflowBehavior =>
+      _rowSpanOverflowBehavior;
+  set rowSpanOverflowBehavior(RowSpanOverflowBehavior value) {
+    if (_rowSpanOverflowBehavior != value) {
+      _rowSpanOverflowBehavior = value;
+      notifyListeners();
+    }
   }
 
   /// Indicates whether the model is sorted.
   ///
   /// The model will be sorted if it has at least one sorted column.
   bool get isSorted =>
-      _columns.firstWhereOrNull((column) => column.sort != null) != null
+      _columns.firstWhereOrNull((column) => column.sortDirection != null) !=
+              null
           ? true
           : false;
 
@@ -104,7 +170,7 @@ class DaviModel<DATA> extends ChangeNotifier {
   bool get isMultiSorted {
     int count = 0;
     for (DaviColumn column in _columns) {
-      if (column.sort != null) {
+      if (column.sortDirection != null) {
         count++;
       }
       if (count > 1) {
@@ -114,7 +180,7 @@ class DaviModel<DATA> extends ChangeNotifier {
     return false;
   }
 
-  DATA rowAt(int index) => _rows[index];
+  DATA rowAt(int index) => _sortableRows[index];
 
   void addRow(DATA row) {
     _originalRows.add(row);
@@ -138,9 +204,10 @@ class DaviModel<DATA> extends ChangeNotifier {
     _updateRows(notify: true);
   }
 
+  /// Removes a row from a given index.
   void removeRowAt(int index) {
     if (_isRowsModifiable) {
-      DATA row = _rows.removeAt(index);
+      DATA row = _sortableRows.removeAt(index);
       _originalRows.remove(row);
     } else {
       _originalRows.removeAt(index);
@@ -148,17 +215,18 @@ class DaviModel<DATA> extends ChangeNotifier {
     notifyListeners();
   }
 
+  /// Removes a row.
   void removeRow(DATA row) {
     _originalRows.remove(row);
     if (_isRowsModifiable) {
-      _rows.remove(row);
+      _sortableRows.remove(row);
     }
     notifyListeners();
   }
 
   DaviColumn<DATA> columnAt(int index) => _columns[index];
 
-  /// Gets a column given an [id]. If [id] is [NULL], no columns are returned.
+  /// Gets a column given an [id]. If [id] is `NULL`, no columns are returned.
   DaviColumn<DATA>? getColumn(dynamic id) {
     if (id != null) {
       for (DaviColumn<DATA> column in _columns) {
@@ -171,23 +239,50 @@ class DaviModel<DATA> extends ChangeNotifier {
   }
 
   void addColumn(DaviColumn<DATA> column) {
-    column.clearSort();
+    if (isSorted && !multiSortEnabled) {
+      DaviColumnHelper.clearSort(column: column);
+    }
     _columns.add(column);
     _checkColumnIdCollision();
+    if (column.sortDirection != null) {
+      _fixSortPriorities();
+    }
+    _checkSummary();
     column.addListener(notifyListeners);
     _ensureSort();
     notifyListeners();
   }
 
+  /// Adds new columns to the model.
   void addColumns(Iterable<DaviColumn<DATA>> columns) {
+    _addColumns(columns, true);
+  }
+
+  void _addColumns(Iterable<DaviColumn<DATA>> columns, bool notify) {
+    final bool sorted = isSorted;
     for (DaviColumn<DATA> column in columns) {
-      column.clearSort();
+      if (sorted && !multiSortEnabled) {
+        DaviColumnHelper.clearSort(column: column);
+      }
       _columns.add(column);
       column.addListener(notifyListeners);
     }
     _checkColumnIdCollision();
+    _fixSortPriorities();
+    _checkSummary();
     _ensureSort();
-    notifyListeners();
+    if (notify) {
+      notifyListeners();
+    }
+  }
+
+  void _fixSortPriorities() {
+    List<DaviColumn<DATA>> sortedColumns = this.sortedColumns;
+    int sortPriority = 1;
+    for (DaviColumn<DATA> column in sortedColumns) {
+      DaviColumnHelper.setSortPriority(
+          column: column, priority: sortPriority++);
+    }
   }
 
   void _checkColumnIdCollision() {
@@ -202,33 +297,32 @@ class DaviModel<DATA> extends ChangeNotifier {
   /// Remove all columns.
   void removeColumns() {
     _columns.clear();
-    _columnInResizing = null;
+    _hasSummary = false;
     _updateRows(notify: true);
   }
 
+  /// Removes a column from a given index.
   void removeColumnAt(int index) {
     DaviColumn<DATA> column = _columns[index];
     removeColumn(column);
   }
 
+  /// Removes a column.
   void removeColumn(DaviColumn<DATA> column) {
     if (_columns.remove(column)) {
       column.removeListener(notifyListeners);
-      if (_columnInResizing == column) {
-        _columnInResizing = null;
-      }
-      if (column.sort != null) {
-        column.clearSort();
-        int priority = 1;
-        for (DaviColumn<DATA> otherColumn in _columns) {
-          if (otherColumn.setSortPriority(priority)) {
-            priority++;
-          }
-        }
+      if (column.sortDirection != null) {
+        DaviColumnHelper.clearSort(column: column);
+        _fixSortPriorities();
         _updateRows(notify: false);
       }
+      _checkSummary();
       notifyListeners();
     }
+  }
+
+  void _checkSummary() {
+    _hasSummary = _columns.any((column) => column.summary != null);
   }
 
   void _notifyOnSort() {
@@ -241,13 +335,13 @@ class DaviModel<DATA> extends ChangeNotifier {
   void clearSort() {
     _clearColumnsSortData();
     _ensureSort();
-    _updateRows(notify: true);
     _notifyOnSort();
+    _updateRows(notify: true);
   }
 
   void _clearColumnsSortData() {
     for (DaviColumn<DATA> column in _columns) {
-      column.clearSort();
+      DaviColumnHelper.clearSort(column: column);
     }
   }
 
@@ -256,6 +350,9 @@ class DaviModel<DATA> extends ChangeNotifier {
   /// If multi sorting is disabled, only the first one in the list will be used.
   /// Not sortable columns will be ignored.
   void sort(List<DaviSort> newSortList) {
+    if (sortingMode == SortingMode.disabled) {
+      return;
+    }
     if (const ListEquality().equals(sortList, newSortList)) {
       // same sort
       return;
@@ -271,7 +368,8 @@ class DaviModel<DATA> extends ChangeNotifier {
       }
       DaviColumn<DATA>? column = getColumn(sort.columnId);
       if (column != null && column.sortable) {
-        column.setSort(sort, priority++);
+        DaviColumnHelper.setSort(
+            column: column, direction: sort.direction, priority: priority++);
         if (!multiSortEnabled) {
           // only the first one
           break;
@@ -279,16 +377,16 @@ class DaviModel<DATA> extends ChangeNotifier {
       }
     }
     _ensureSort();
-    _updateRows(notify: true);
     _notifyOnSort();
+    _updateRows(notify: true);
   }
 
   void _ensureSort() {
-    if (alwaysSorted) {
+    if (sortingMode == SortingMode.alwaysSorted) {
       DaviColumn? firstNonSortedColumn;
       for (DaviColumn column in _columns) {
         if (column.sortable) {
-          if (column.sort != null) {
+          if (column.sortDirection != null) {
             // It's already sorted.
             return;
           }
@@ -297,8 +395,10 @@ class DaviModel<DATA> extends ChangeNotifier {
       }
       // No sorted columns. Let's order the first one available.
       if (firstNonSortedColumn != null) {
-        firstNonSortedColumn.setSort(
-            DaviSort(firstNonSortedColumn.id, DaviSortDirection.ascending), 1);
+        DaviColumnHelper.setSort(
+            column: firstNonSortedColumn,
+            direction: DaviSortDirection.ascending,
+            priority: 1);
       }
     }
   }
@@ -312,28 +412,61 @@ class DaviModel<DATA> extends ChangeNotifier {
   void _updateRows({required bool notify}) {
     if (isSorted && !ignoreDataComparators) {
       List<DATA> list = List.from(_originalRows);
-      list.sort(_compoundSort);
-      _rows = list;
+
+      // Create a list of pairs (index, value),
+      // where each element is a MapEntry
+      // The key of MapEntry is the index,
+      // and the value is the corresponding element
+      List<MapEntry<int, DATA>> indexedList = List.generate(
+        list.length,
+        (index) => MapEntry(index, list[index]),
+      );
+
+      indexedList.sort((a, b) => _compoundSort(a.value, a.key, b.value, b.key));
+
+      // Convert the sorted indexedList back into a normal list of values
+      // This gives us the sorted list of values without the indices
+      _sortableRows = indexedList.map((entry) => entry.value).toList();
     } else {
-      _rows = UnmodifiableListView(_originalRows);
+      _sortableRows = UnmodifiableListView(_originalRows);
     }
+    _rowsView = UnmodifiableListView(_sortableRows);
     if (notify) {
       notifyListeners();
     }
   }
 
   /// Function to realize the multi sort.
-  int _compoundSort(DATA a, DATA b) {
+  int _compoundSort(DATA rowA, int indexA, DATA rowB, int indexB) {
     int r = 0;
     for (final DaviColumn<DATA> column in sortedColumns) {
-      if (column.sort != null) {
-        final DaviDataComparator<DATA> dataComparator = column.dataComparator;
-        final DaviSortDirection direction = column.sort!.direction;
-
+      final DaviSortDirection? direction = column.sortDirection;
+      if (direction != null) {
+        final DaviComparator<DATA> dataComparator = column.dataComparator;
+        ValueMapperParams<DATA>? valueParams;
+        BarValueMapperParams<DATA>? barValueParams;
+        IconMapperParams<DATA>? iconParams;
+        dynamic cellValueA, cellValueB;
+        if (column.cellValue != null) {
+          cellValueA = column.cellValue!(DaviColumnHelper.valueParams(
+              params: valueParams, data: rowA, rowIndex: indexA));
+          cellValueB = column.cellValue!(DaviColumnHelper.valueParams(
+              params: valueParams, data: rowB, rowIndex: indexB));
+        } else if (column.cellBarValue != null) {
+          cellValueA = column.cellBarValue!(DaviColumnHelper.barValueParams(
+              params: barValueParams, data: rowA, rowIndex: indexA));
+          cellValueB = column.cellBarValue!(DaviColumnHelper.barValueParams(
+              params: barValueParams, data: rowB, rowIndex: indexB));
+        } else if (column.cellIcon != null) {
+          cellValueA = column.cellIcon!(DaviColumnHelper.iconParams(
+              params: iconParams, data: rowA, rowIndex: indexA));
+          cellValueB = column.cellIcon!(DaviColumnHelper.iconParams(
+              params: iconParams, data: rowB, rowIndex: indexB));
+        }
         if (direction == DaviSortDirection.descending) {
-          r = dataComparator(b, a, column);
+          r = dataComparator(cellValueB, cellValueA, rowB, rowA);
         } else {
-          r = dataComparator(a, b, column);
+          r = dataComparator(cellValueA, cellValueB, rowA, rowB);
         }
         if (r != 0) {
           break;
