@@ -1,5 +1,9 @@
 import 'package:davi/src/column.dart';
 
-/// Signature to the callback of when a sort has occurred.
+/// This callback is triggered before the internal sorting logic is executed.
+///
+/// Will only be triggered when the header is clicked or when
+/// the [DaviModel.sort] or [DaviModel.clearSort] methods
+/// are called programmatically.
 typedef OnSortCallback<DATA> = void Function(
     List<DaviColumn<DATA>> sortedColumns);
